Refactor of identity package
- Clearer grouping in files
- Improve test coverage
