+++ pkg/identity/sign.go
@@ -10,8 +10,6 @@ import (
 	"crypto"
 	"crypto/ecdsa"
 	"crypto/rand"
-	"crypto/x509"
-	"encoding/pem"
 	"fmt"
 
 	"github.com/hyperledger/fabric/bccsp/utils"
@@ -44,33 +42,4 @@ func ecdsaPrivateKeySign(privateKey *ecdsa.PrivateKey) Sign {
 
 		return utils.MarshalECDSASignature(r, s)
 	}
-}
-
-// PrivateKeyFromPEM creates a private key from PEM encoded data.
-func PrivateKeyFromPEM(privateKeyPEM []byte) (crypto.PrivateKey, error) {
-	block, _ := pem.Decode(privateKeyPEM)
-	if block == nil {
-		return nil, fmt.Errorf("Failed to parse private key PEM")
-	}
-
-	privateKey, err := x509.ParsePKCS8PrivateKey(block.Bytes)
-	if err != nil {
-		return nil, err
-	}
-
-	return privateKey, nil
-}
-
-// PrivateKeyToPEM converts a private key to PEM encoded PKCS #8 data.
-func PrivateKeyToPEM(privateKey crypto.PrivateKey) ([]byte, error) {
-	privateKeyBytes, err := x509.MarshalPKCS8PrivateKey(privateKey)
-	if err != nil {
-		return nil, err
-	}
-
-	block := &pem.Block{
-		Type:  "PRIVATE KEY",
-		Bytes: privateKeyBytes,
-	}
-	return pemEncode(block)
 }
